+++ NEWS.md
@@ -4,8 +4,9 @@
 * Allow `btb_ptsToGrid` to use simple feature points as input
 * Change `kernelSmoothing` name (to `btb_smooth`)
 * Allow `btb_ptsToGrid` to use simple feature points as input
+* New function to link points to their centroid in a grid (`btb_add_centroides`)
 * Old functions `dfToGrid` and `kernelSmoothing` replaced by warnings
-* Possibility to use Inspire naming for grid cells (with `add_inspire` function) 
+* Possibility to use Inspire naming for grid cells (with `btb_add_inspire` function) 
 * Roxygen documentation
 * testthat compatible tests
 * New tests